breaks out Game.tsx file components
