+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import './App.css';
-import PaiSho from "./components/Game";
+import Game from "./components/Game";
 
 
 function App(): JSX.Element{
   return (
       <div className="App">
-        <PaiSho></PaiSho>
+          <Game height={300} width={300} radius={300} gridCount={18}></Game>
       </div>
   );
 }
